Protected Routed Logged in user to get contacts

+++ controllers/contactController.js
@@ -7,16 +7,16 @@ const Contact = require("../models/contactModal");
 
 //@desc Get all Contacts
 //@route GET /api/contacts
-//@access public
+//@access private
 
 const getContacts = asyncHandler(async (req, res) => {
-  const contacts = await Contact.find();
+  const contacts = await Contact.find({ user_id: req.user.id });
   res.status(200).json(contacts);
 });
 
 //@desc Create New Contacts
 //@route POST /api/contacts
-//@access public
+//@access private
 
 const createContact = asyncHandler(async (req, res) => {
   console.log("The request body is", req.body);
@@ -30,13 +30,14 @@ const createContact = asyncHandler(async (req, res) => {
     name,
     email,
     phone,
+    user_id: req.user.id,
   });
   res.status(201).json(contact);
 });
 
 //@desc Get Contact
 //@route POST /api/contacts/:id
-//@access public
+//@access private
 
 const getContact = asyncHandler(async (req, res) => {
   const contact = await Contact.findById(req.params.id);
@@ -49,7 +50,7 @@ const getContact = asyncHandler(async (req, res) => {
 
 //@desc Update Contact
 //@route PUT /api/contacts/:id
-//@access public
+//@access private
 
 const updateContact = asyncHandler(async (req, res) => {
   const contact = await Contact.findById(req.params.id);
@@ -57,6 +58,12 @@ const updateContact = asyncHandler(async (req, res) => {
     res.status(404);
     throw new Error("Contact not found");
   }
+
+  if (contact.user_id.toString() !== req.user.id) {
+    res.status(403);
+    throw new Error("user don't have permission to update other user contacts");
+  }
+
   const updatedContact = await Contact.findByIdAndUpdate(
     req.params.id,
     req.body,
@@ -68,7 +75,7 @@ const updateContact = asyncHandler(async (req, res) => {
 
 //@desc Delete Contacts
 //@route DELETE /api/contacts/:id
-//@access public
+//@access private
 
 const deleteContact = asyncHandler(async (req, res) => {
   const contact = await Contact.findById(req.params.id);
@@ -76,10 +83,10 @@ const deleteContact = asyncHandler(async (req, res) => {
     res.status(404);
     throw new Error("Contact not found");
   }
-  // if (contact.user_id.toString() !== req.user.id) {
-  //   res.status(403);
-  //   throw new Error("User don't have permission to update other user contacts");
-  // }
+  if (contact.user_id.toString() !== req.user.id) {
+    res.status(403);
+    throw new Error("User don't have permission to delete other user contacts");
+  }
   await Contact.deleteOne({ _id: req.params.id });
   res.status(200).json(contact);
 });
